requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0

# Speech recognition (GPU-accelerated)
faster-whisper>=1.1.0

# Text-to-Speech: standalone Piper binary (NOT piper-tts Python package)
# The Python package has broken phonemizer in Docker — standalone binary works

# Voice Activity Detection
onnxruntime>=1.16.0
numpy>=1.24.0

# Web dashboard
flask>=3.0.0
flask-socketio>=5.3.0

# System monitoring
psutil>=5.9.0

# Server monitoring
icmplib>=3.0.4

# Utilities
pathvalidate>=2.5.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
